Deleted unnecesssary files and removed commented out code
