+++ ludwig/schema/__init__.py
@@ -22,8 +22,6 @@ from jsonschema.validators import extend
 
 from ludwig.schema.trainer import get_model_type_jsonschema, get_trainer_jsonschema
 from ludwig.constants import MODEL_TYPE, INPUT_FEATURES, OUTPUT_FEATURES, COMBINER, TRAINER, PREPROCESSING, HYPEROPT
-from ludwig.decoders.registry import get_decoder_classes
-from ludwig.encoders.registry import get_encoder_classes
 from ludwig.schema.features.utils import get_input_feature_jsonschema, get_output_feature_jsonschema
 from ludwig.schema.combiners.utils import get_combiner_jsonschema
 from ludwig.schema.trainer import get_trainer_jsonschema
@@ -31,8 +29,6 @@ from ludwig.schema.utils import create_cond
 
 
 def get_schema():
-    from ludwig.combiners.combiners import get_combiner_schema
-
     schema = {
         "type": "object",
         "properties": {
