+++ ludwig/schema/features/bag_feature.py
@@ -3,12 +3,11 @@ from typing import Optional
 from marshmallow_dataclass import dataclass
 
 from ludwig.schema import utils as schema_utils
-from ludwig.schema.features import base
-from ludwig.schema.features.preprocessing import BasePreprocessingConfig, PreprocessingDataclassField
+from ludwig.schema.preprocessing import BasePreprocessingConfig, PreprocessingDataclassField
 
 
 @dataclass
-class BagInputFeatureConfig(base.BaseFeatureConfig):
+class BagInputFeatureConfig(schema_utils.BaseMarshmallowConfig):
     """
     BagInputFeatureConfig is a dataclass that configures the parameters used for a bag input feature.
     """
