+++ ludwig/schema/features/sequence_feature.py
@@ -3,12 +3,11 @@ from typing import Optional
 from marshmallow_dataclass import dataclass
 
 from ludwig.schema import utils as schema_utils
-from ludwig.schema.features import base
-from ludwig.schema.features.preprocessing import BasePreprocessingConfig, PreprocessingDataclassField
+from ludwig.schema.preprocessing import BasePreprocessingConfig, PreprocessingDataclassField
 
 
 @dataclass
-class SequenceInputFeatureConfig(base.BaseFeatureConfig):
+class SequenceInputFeatureConfig(schema_utils.BaseMarshmallowConfig):
     """
     SequenceInputFeatureConfig is a dataclass that configures the parameters used for a sequence input feature.
     """
@@ -33,7 +32,7 @@ class SequenceInputFeatureConfig(base.BaseFeatureConfig):
 
 
 @dataclass
-class SequenceOutputFeatureConfig(base.BaseFeatureConfig):
+class SequenceOutputFeatureConfig(schema_utils.BaseMarshmallowConfig):
     """
     SequenceOutputFeatureConfig is a dataclass that configures the parameters used for a sequence output feature.
     """
