+++ ludwig/schema/features/set_feature.py
@@ -3,12 +3,11 @@ from typing import Optional
 from marshmallow_dataclass import dataclass
 
 from ludwig.schema import utils as schema_utils
-from ludwig.schema.features import base
-from ludwig.schema.features.preprocessing import BasePreprocessingConfig, PreprocessingDataclassField
+from ludwig.schema.preprocessing import BasePreprocessingConfig, PreprocessingDataclassField
 
 
 @dataclass
-class SetInputFeatureConfig(base.BaseFeatureConfig):
+class SetInputFeatureConfig(schema_utils.BaseMarshmallowConfig):
     """
     SetInputFeatureConfig is a dataclass that configures the parameters used for a set input feature.
     """
@@ -33,7 +32,7 @@ class SetInputFeatureConfig(base.BaseFeatureConfig):
 
 
 @dataclass
-class SetOutputFeatureConfig(base.BaseFeatureConfig):
+class SetOutputFeatureConfig(schema_utils.BaseMarshmallowConfig):
     """
     SetOutputFeatureConfig is a dataclass that configures the parameters used for a set output feature.
     """
