+++ ludwig/schema/features/text_feature.py
@@ -3,12 +3,11 @@ from typing import Optional
 from marshmallow_dataclass import dataclass
 
 from ludwig.schema import utils as schema_utils
-from ludwig.schema.features import base
-from ludwig.schema.features.preprocessing import BasePreprocessingConfig, PreprocessingDataclassField
+from ludwig.schema.preprocessing import BasePreprocessingConfig, PreprocessingDataclassField
 
 
 @dataclass
-class TextInputFeatureConfig(base.BaseFeatureConfig):
+class TextInputFeatureConfig(schema_utils.BaseMarshmallowConfig):
     """
     TextInputFeatureConfig is a dataclass that configures the parameters used for a text input feature.
     """
@@ -35,7 +34,7 @@ class TextInputFeatureConfig(base.BaseFeatureConfig):
 
 
 @dataclass
-class TextOutputFeatureConfig(base.BaseFeatureConfig):
+class TextOutputFeatureConfig(schema_utils.BaseMarshmallowConfig):
     """
     TextOutputFeatureConfig is a dataclass that configures the parameters used for a text output feature.
     """
