+++ ludwig/schema/features/utils.py
@@ -1,9 +1,6 @@
 from ludwig.utils.registry import Registry
 from ludwig.schema import utils as schema_utils
 
-from ludwig.decoders.registry import get_decoder_classes
-from ludwig.encoders.registry import get_encoder_classes
-
 input_type_registry = Registry()
 output_type_registry = Registry()
 
@@ -109,109 +106,3 @@ def get_output_feature_conds():
         feature_cond = schema_utils.create_cond({"type": feature_type}, feature_props)
         conds.append(feature_cond)
     return conds
-
-
-# def get_encoder_conds(input_feature_types):
-#     conds = []
-#     for feature_type in input_feature_types:
-#         encoder_names = list(get_encoder_classes(feature_type).keys())
-#         encoder_cond = create_cond(
-#             {"type": feature_type},
-#             {"encoder": {"enum": encoder_names}},
-#         )
-#         conds.append(encoder_cond)
-#     return conds
-#
-#
-# def get_decoder_conds(input_feature_types):
-#     conds = []
-#     for feature_type in input_feature_types:
-#         decoder_names = list(get_decoder_classes(feature_type).keys())
-#         decoder_cond = create_cond(
-#             {"type": feature_type},
-#             {"decoder": {"enum": decoder_names}},
-#         )
-#         conds.append(decoder_cond)
-#     return conds
-#
-#
-# def get_input_preproc_conds(input_feature_types):
-#     conds = []
-#     for feature_type in input_feature_types:
-#         feature_cls = input_type_registry[feature_type]
-#         preproc_spec = {
-#             "type": "object",
-#             "properties": feature_cls.preprocessing_schema(),
-#             "additionalProperties": False,
-#         }
-#         preproc_cond = create_cond(
-#             {"type": feature_type},
-#             {"preprocessing": preproc_spec},
-#         )
-#         conds.append(preproc_cond)
-#     return conds
-#
-#
-# def get_output_preproc_conds(output_feature_types):
-#     conds = []
-#     for feature_type in output_feature_types:
-#         feature_cls = output_type_registry[feature_type]
-#         preproc_spec = {
-#             "type": "object",
-#             "properties": feature_cls.preprocessing_schema(),
-#             "additionalProperties": False,
-#         }
-#         preproc_cond = create_cond(
-#             {"type": feature_type},
-#             {"preprocessing": preproc_spec},
-#         )
-#         conds.append(preproc_cond)
-#     return conds
-
-
-# def get_input_feature_jsonschema():
-#     """
-#     This function returns a JSON schema structured to only requires a `type` key and then conditionally applies a
-#     corresponding input feature's field constraints.
-#
-#     Returns: JSON Schema
-#
-#     """
-#     input_feature_types = sorted(list(input_type_registry.keys()))
-#     return {
-#         "type": "array",
-#         "items": {
-#             "type": "object",
-#             "properties": {
-#                 "name": {"type": "string"},
-#                 "type": {"type": "string", "enum": input_feature_types},
-#                 "column": {"type": "string"},
-#                 "encoder": {"type": "string"},
-#             },
-#             "allOf": get_encoder_conds(input_feature_types) + get_input_preproc_conds(input_feature_types),
-#             "required": ["name", "type"],
-#         }
-#     }
-
-# def get_output_feature_jsonschema():
-#     """
-#     This function returns a JSON schema structured to only requires a `type` key and then conditionally applies a
-#     corresponding output feature's field constraints.
-#
-#     Returns: JSON Schema
-#     """
-#     output_feature_types = sorted(list(output_type_registry.keys()))
-#     return {
-#         "type": "array",
-#         "items": {
-#             "type": "object",
-#             "properties": {
-#                 "name": {"type": "string"},
-#                 "type": {"type": "string", "enum": output_feature_types},
-#                 "column": {"type": "string"},
-#                 "decoder": {"type": "string"},
-#             },
-#             "allOf": get_decoder_conds(output_feature_types) + get_output_preproc_conds(output_feature_types),
-#             "required": ["name", "type"],
-#         },
-#     }
